rtl8189es: Make module auto-load
Add a proper sdio: module-alias and move the initial powerup + rescan to
rtl8189es_wifi_gpio_init, so that the wifi module gets auto-loaded if
enabled in the fex file. This assumes that mmc_pm is build into the kernel,
as mmc_pm is the one doing the sdio_wifi_para fex section parsing.

If mmc_pm is not buildin it will need to be loaded manually (just as before),
and then on mmc_pm load rtl8189es will be autoloaded.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>