sunxi-sound: Move registration of hdmi-audio devices to sunxi-hdmi code
Registering the hdmi audio-devices when not using hdmi means that userspace
may end up trying to use hdmi while it is not working. When hdmi-audio is
not enabled, and userspace still tries to use it the audio-using app(s) will
hang when trying to write to the audio-device, since the write will never
complete when the video-part is not consuming any hdmi-audio.

Note that this patch does not unregister the audio-card again if the hdmi
block is powered down (dpms / blank screen), so as to avoid audio-streaming
apps to crash / abort because the soundcard has gone away. Instead the app
will pause (write will block) on monitor powerdown and resume again when
the monitor powers back on.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>

Backported to 3.0
Signed-off-by: Oliver Schinagl <oliver@schinagl.nl>

Conflicts:
	sound/soc/sun4i/hdmiaudio/sun4i-hdmiaudio.c
	sound/soc/sun4i/hdmiaudio/sun4i-hdmipcm.c